# lib/api-zod/src/generated/types/crmRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI CSV Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface CrmRecord {
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  country_code?: string | null;
  /** @nullable */
  mobile_without_country_code?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  lead_owner?: string | null;
  /** @nullable */
  crm_status?: string | null;
  /** @nullable */
  crm_note?: string | null;
  /** @nullable */
  data_source?: string | null;
  /** @nullable */
  possession_time?: string | null;
  /** @nullable */
  description?: string | null;
}