# lib/api-zod/src/generated/types/importCSVInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { ImportCSVInputRowsItem } from './importCSVInputRowsItem';

export interface ImportCSVInput {
  /** Parsed CSV rows to process through AI */
  rows: ImportCSVInputRowsItem[];
  /** Column headers for AI context */
  columns: string[];
}