# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { CrmRecord } from './crmRecord';

export interface ImportResult {
  records: CrmRecord[];
  imported: number;
  skipped: number;
  total: number;
}