# lib/api-zod/src/generated/types/parseCSVInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI CSV Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface ParseCSVInput {
  /** Raw CSV file contents as a string */
  csvText: string;
}