# lib/api-zod/src/generated/types/parsedCSV.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedCSVRowsItem } from './parsedCSVRowsItem';

export interface ParsedCSV {
  /** Column headers from the CSV */
  columns: string[];
  /** Parsed rows as key-value pairs */
  rows: ParsedCSVRowsItem[];
  /** Total number of data rows parsed */
  totalRows: number;
}